Move PRIORITY to new type directory
