+++ tests/test_types.py
@@ -3,10 +3,7 @@
 import datetime
 from typing import Optional, Union
 
-import pytest
-from pydantic import ValidationError
-
-from ical._types import ICS_ENCODERS, ComponentModel, Priority
+from ical._types import ICS_ENCODERS, ComponentModel
 from ical.parsing.component import ParsedComponent
 from ical.parsing.property import ParsedProperty
 
@@ -136,24 +133,3 @@ def test_optional_field_parser() -> None:
         {"dt": [ParsedProperty(name="dt", value="20220724T120000")]}
     )
     assert model.dt == datetime.datetime(2022, 7, 24, 12, 0, 0)
-
-
-def test_priority() -> None:
-    """Test for priority fields."""
-
-    class TestModel(ComponentModel):
-        """Model under test."""
-
-        pri: Priority
-
-    model = TestModel.parse_obj({"pri": [ParsedProperty(name="dt", value="1")]})
-    assert model.pri == 1
-
-    model = TestModel.parse_obj({"pri": [ParsedProperty(name="dt", value="9")]})
-    assert model.pri == 9
-
-    with pytest.raises(ValidationError):
-        TestModel.parse_obj({"pri": [ParsedProperty(name="dt", value="-1")]})
-
-    with pytest.raises(ValidationError):
-        TestModel.parse_obj({"pri": [ParsedProperty(name="dt", value="10")]})
